Store counted strings in StringPool

+++ src/include/ConfigDB/Store.h
@@ -82,7 +82,7 @@ public:
 	void clear();
 
 	String getValueString(const PropertyInfo& info, const void* data) const;
-	PropertyData parseString(const PropertyInfo& prop, const char* value, size_t valueLength);
+	PropertyData parseString(const PropertyInfo& prop, const char* value, uint16_t valueLength);
 
 	const StringPool& getStringPool() const
 	{
